StorageParser package
